+++ Changelog.md
@@ -24,6 +24,7 @@ Bugfixes:
  * Type Checker: Disallow signed literals as exponent in exponentiation operator.
  * Allow `type(Contract).name` for abstract contracts and interfaces.
  * Type Checker: Disallow structs containing nested mapping in memory as parameters for library functions.
+ * Type Checker: Disallow ``using for`` directive inside interfaces.
 
 
 ### 0.7.0 (2020-07-28)